net-snmp: fix ASN_INTEGER scalar value size
Use long storage for ASN_INTEGER scalar values passed to snmp_set_var_typed_value().  PPTP, L2TP, and PPPoE starting/active handlers previously passed unsigned int locals, and statCoreCPU passed triton_stat.cpu directly, which does not match Net-SNMP's C representation for ASN_INTEGER on 64-bit systems.

Keep the exposed MIB values and access paths unchanged; only stage the values through correctly sized local variables before encoding them for Net-SNMP.

+++ accel-pppd/extra/net-snmp/statPPPOE.c
@@ -71,7 +71,7 @@ static int handle_statPPPOEStarting(netsnmp_mib_handler *handler,
 				    netsnmp_agent_request_info *reqinfo,
 				    netsnmp_request_info *requests)
 {
-	unsigned int stat;
+	long stat;
 
 	switch (reqinfo->mode) {
 	case MODE_GET:
@@ -92,7 +92,7 @@ static int handle_statPPPOEActive(netsnmp_mib_handler *handler,
 				  netsnmp_agent_request_info *reqinfo,
 				  netsnmp_request_info *requests)
 {
-	unsigned int stat;
+	long stat;
 
 	switch (reqinfo->mode) {
 	case MODE_GET:
